WA-89 Moved DB stuff to login-core.js
